services/user_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.2
pydantic-settings==2.1.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
greenlet==3.0.1
alembic==1.13.1
redis==5.0.1
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0

# Dev dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
aiosqlite==0.19.0